Webpacker works, but now it's time for react
Travis requires some hand-holding to get this build running now. It's a
`ruby` project, so it doesn't come with yarn, so we have to go get that
and then make sure that we `webpack` before running the tests.

Also, I've left a silly react auto-generated component in here so we can
start from scratch on the next PR but still have a template (it's been a
while, I need some hand-holding).